requirements.txt
absl-py==1.4.0
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
contourpy==1.0.7
cryptography==41.0.1
cycler==0.11.0
filelock==3.12.0
fonttools==4.39.3
gdown==4.7.1
google-auth==2.20.0
google-auth-oauthlib==1.0.0
grpcio==1.56.0
h5py==3.8.0
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
joblib==1.2.0
kiwisolver==1.4.4
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
numpy==1.23.5
oauthlib==3.2.2
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.2.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.29.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==0.22.1
scipy==1.10.1
signatory==1.2.6.1.9.0
six==1.16.0
soupsieve==2.4.1
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboardX==2.6
threadpoolctl==3.1.0
torch==1.8.1+cu101
torchaudio==0.8.1
torchvision==0.9.1+cu101
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.3.6
zipp==3.15.0
